[nanvix-bench] E: Split into modules
Extract each benchmark implementation from main.rs into its own
file under a new benchmarks/ module directory. Shared constants,
helpers (prepare_new_message, setup, start, kill, cleanup,
run_user_vm_echo_once), and infrastructure (start_nanvixd,
start_user_vm) move into benchmarks/mod.rs.

No behavioral changes.
