+++ src/utils/nanvix-bench/src/main.rs
@@ -17,108 +17,32 @@
 
 mod args;
 mod benchmark;
+mod benchmarks;
 
 //==================================================================================================
 // Imports
 //==================================================================================================
 
+#[cfg(not(feature = "timestamp-messages"))]
+use crate::benchmark::UserVmDeployment;
 use crate::{
     args::Args,
     benchmark::{
         Benchmark,
         BenchmarkFlavour,
         LinuxdDeployment,
-        UserVmDeployment,
     },
 };
 use ::anyhow::Result;
-use ::indicatif::{
-    ProgressBar,
-    ProgressStyle,
-};
-use ::log::{
-    debug,
-    error,
-    warn,
-};
+use ::log::error;
 use ::nanvix::{
-    http::{
-        message,
-        message::{
-            ErrorResponse,
-            HTTP_HEADER_MESSAGE_TYPE,
-            Kill,
-            KillResponse,
-            MessageType,
-            New,
-        },
-    },
     hwloc,
     hwloc::HwLoc,
-    sandbox::UserVmIdentifier,
-    sys::{
-        ipc::{
-            DataChunk,
-            DataChunkHeader,
-            IkcFrame,
-            Message,
-        },
-        pm::ThreadIdentifier,
-    },
-    syscall::{
-        LinuxDaemonMessage,
-        unistd::message::{
-            ReadRequest,
-            ReadResponse,
-            WriteResponse,
-        },
-    },
-    syscomm::{
-        ReadExact,
-        SocketStream,
-        SocketType,
-        UnboundSocket,
-        WriteAll,
-    },
-    uservm,
-    uservm::{
-        UserVm,
-        UserVmArgs,
-        counters::MessageCounters,
-        orchestrator::{
-            IoControlCommand,
-            IoControlResponse,
-        },
-    },
-};
-use ::reqwest::header::{
-    CONTENT_TYPE,
-    HeaderMap,
 };
 use ::std::{
-    collections::HashMap,
     fs::File,
     io::BufReader,
-    mem,
-    net::TcpStream,
-    process::{
-        self,
-        Child,
-        Command,
-        Stdio,
-    },
-    time::{
-        Duration,
-        Instant,
-    },
-};
-use ::tokio::{
-    sync::mpsc,
-    task::JoinHandle,
-    time::{
-        sleep,
-        timeout,
-    },
+    time::Duration,
 };
 
 //==================================================================================================
@@ -137,1533 +61,13 @@ const CARGO_PKG_NAME: &str = match option_env!("CARGO_PKG_NAME") {
 ///
 /// # Description
 ///
-/// Sleep duration (in ms) to wait for the system to clean up after a benchmark run.
-///
-const CLEANUP_SLEEP_DURATION: u64 = 10;
-
-///
-/// # Description
-///
-/// Sleep duration (in ms) after the warmup echo, before timed iterations begin.
-///
-const WARMUP_SLEEP_DURATION: u64 = CLEANUP_SLEEP_DURATION;
-
-///
-/// # Description
-///
 /// Timeout (in seconds) for HTTP requests to nanvixd (start, kill, etc.).
 ///
 const NANVIXD_HTTP_TIMEOUT_SECS: u64 = 60;
 
-///
-/// # Description
-///
-/// Timeout (in seconds) to wait for nanvixd to exit after SIGINT before sending SIGKILL.
-///
-const NANVIXD_SHUTDOWN_TIMEOUT_SECS: u64 = 30;
-
-///
-/// # Description
-///
-/// Sleep duration (in ms) to wait for the system to clean up after a benchmark run when deploying
-/// linuxd in an L2 VM. We need a longer clean-up for cloud-hypervisor to shutdown.
-///
-const CLEANUP_L2_SLEEP_DURATION: u64 = 100;
-
-///
-/// # Description
-///
-/// Timeout (in seconds) for the gateway connection retry loop. Matches the nanvixd-side gateway
-/// probe timeout (`GATEWAY_CONNECT_TIMEOUT`).
-///
-const GATEWAY_CONNECT_TIMEOUT_SECS: u64 = 60;
-
-///
-/// # Description
-///
-/// Timeout (in seconds) for echo I/O operations (write + read) on the gateway stream.
-///
-const ECHO_IO_TIMEOUT_SECS: u64 = 60;
-
-///
-/// # Description
-///
-/// Sleep duration (in milliseconds) between gateway connection retry attempts.
-///
-const GATEWAY_CONNECT_RETRY_SLEEP_MS: u64 = 1;
-
-///
-/// # Description
-///
-/// Maximum number messages that can be queued in a channel.
-///
-pub const CHANNEL_CAPACITY: usize = 1024;
-
-const NANVIXD_ADDRESS: &str = "127.0.0.1:9999";
-
-///
-/// # Description
-///
-/// Default size of the message we are sending to the user VM.
-const DEFAULT_PAYLOAD_SIZE: usize = 32;
-
-const DEFAULT_TENANT_ID: &str = "foo";
-const DEFAULT_APP_NAME: &str = "bar";
-
 //==================================================================================================
-// Implementations
+// Main
 //==================================================================================================
-
-impl Benchmark {
-    fn prepare_new_message(
-        &self,
-        tenant_id: Option<String>,
-        app_name: Option<String>,
-    ) -> Result<(HeaderMap, message::New)> {
-        let mut new_msg_headers = HeaderMap::new();
-        new_msg_headers.insert(CONTENT_TYPE, "application/json".parse()?);
-        new_msg_headers
-            .insert(HTTP_HEADER_MESSAGE_TYPE, format!("{}", message::MessageType::New).parse()?);
-
-        let new_msg = message::New {
-            tenant_id: tenant_id.unwrap_or(DEFAULT_TENANT_ID.to_string()),
-            app_name: app_name.unwrap_or(DEFAULT_APP_NAME.to_string()),
-            program: self.flavour.get_program(&self.workspace_root),
-            program_args: "".to_string(),
-        };
-
-        Ok((new_msg_headers, new_msg))
-    }
-
-    /// Start nanvixd and, optionally, configure it to deploy linuxd inside an L2 VM.
-    fn start_nanvixd(&self, linuxd_deployment: &LinuxdDeployment) -> Result<Child> {
-        let mut nanvixd_args: Vec<String> = vec![
-            format!("{}/bin/nanvixd.elf", self.workspace_root.display()),
-            ::nanvixd::args::Args::OPT_HTTP_SOCKADDR.to_string(),
-            NANVIXD_ADDRESS.to_string(),
-            ::nanvixd::args::Args::OPT_TOOLCHAIN_BIN_DIRECTORY.to_string(),
-            self.nanvixd_toolchain_bin_dir.clone(),
-            ::nanvixd::args::Args::OPT_TMP_DIRECTORY.to_string(),
-            self.nanvixd_tmp_dir.clone(),
-        ];
-        if let Some(netns_pool_size) = self.nanvixd_netns_pool_size {
-            nanvixd_args.push(::nanvixd::args::Args::OPT_NETNS_POOL_SIZE.to_string());
-            nanvixd_args.push(netns_pool_size.to_string());
-        }
-        if let Some(hwloc_file) = &self.hwloc_file {
-            nanvixd_args.push(::nanvixd::args::Args::OPT_HWLOC.to_string());
-            nanvixd_args.push(hwloc_file.clone());
-        }
-        if *linuxd_deployment == LinuxdDeployment::L2Vm {
-            nanvixd_args.push(::nanvixd::args::Args::OPT_L2.to_string());
-        }
-
-        debug!("Starting nanvixd with command: {}", nanvixd_args.join(" "));
-        let nanvixd_cmd = Command::new(&nanvixd_args[0])
-            .args(&nanvixd_args[1..])
-            .stdin(Stdio::null())
-            .stdout(Stdio::piped())
-            .current_dir(&self.workspace_root)
-            .spawn()?;
-
-        Ok(nanvixd_cmd)
-    }
-
-    /// Auxiliary method to start a user VM used by low-level benchmarks that want to bypass
-    /// nanvixd.
-    fn start_user_vm(&self, gateway_addr: Option<String>) -> Result<Child> {
-        let mut user_vm_args: Vec<String> = vec![
-            format!("{}/bin/uservm.elf", self.workspace_root.display()),
-            uservm::args::Args::OPT_USER_VM_ID.to_string(),
-            "1".to_string(),
-            uservm::args::Args::OPT_KERNEL.to_string(),
-            format!("{}/bin/kernel.elf", self.workspace_root.display()),
-            uservm::args::Args::OPT_INITRD.to_string(),
-            self.flavour.get_program(&self.workspace_root),
-        ];
-        if let Some(gateway_addr) = gateway_addr {
-            user_vm_args.push(uservm::args::Args::OPT_SYSTEM_VM_SOCKADDR.to_string());
-            user_vm_args.push(gateway_addr);
-        }
-        if let Some(hwloc) = self.hwloc.clone() {
-            let taskset: Vec<String> = vec![
-                "taskset".to_string(),
-                "-ac".to_string(),
-                hwloc.get_nanovm_core_str(),
-            ];
-            user_vm_args.splice(0..0, taskset);
-        }
-
-        debug!("Starting user VM with command: {}", user_vm_args.join(" "));
-        let user_vm_cmd = Command::new(&user_vm_args[0])
-            .args(&user_vm_args[1..])
-            .stdin(Stdio::null())
-            .stdout(Stdio::null())
-            .current_dir(&self.workspace_root)
-            .spawn()?;
-
-        Ok(user_vm_cmd)
-    }
-
-    /// Configures the set-up by starting linuxd and the gateway server.
-    pub fn setup(&mut self, linuxd_deployment: &LinuxdDeployment) {
-        match self.start_nanvixd(linuxd_deployment) {
-            Ok(nanvixd) => self.nanvixd = Some(nanvixd),
-            Err(_) => {
-                error!("error starting up nanvixd");
-                self.cleanup();
-                process::exit(1);
-            },
-        }
-
-        while TcpStream::connect_timeout(
-            &NANVIXD_ADDRESS.to_string().parse().unwrap(),
-            Duration::from_millis(10),
-        )
-        .is_err()
-        {
-            continue;
-        }
-
-        debug!("nanvixd is ready to serve requests");
-    }
-
-    /// Starts the Nano VM via POST request to nanvixd. Returns the user VM ID as well as an open
-    /// socket to interact with the VMs stdin/stdout.
-    pub async fn start(
-        &mut self,
-        payload: message::New,
-        headers: HeaderMap,
-        linuxd_deployment: &LinuxdDeployment,
-    ) -> Result<(UserVmIdentifier, SocketStream)> {
-        let http_response: ::reqwest::Response = self
-            .nanvixd_client
-            .post(format!("http://{}", NANVIXD_ADDRESS))
-            .headers(headers)
-            .json(&payload)
-            .send()
-            .await?;
-
-        // Check if the response is successful
-        let status: ::reqwest::StatusCode = http_response.status();
-        if !status.is_success() {
-            // Try to deserialize as ErrorResponse to get detailed error info
-            let error_msg: String = match http_response.json::<ErrorResponse>().await {
-                Ok(err_response) => {
-                    format!(
-                        "nanvixd returned error (status={}, code={:?}): {}",
-                        status, err_response.code, err_response.message
-                    )
-                },
-                Err(e) => {
-                    format!(
-                        "nanvixd returned error (status={}): failed to parse error response: {}",
-                        status, e
-                    )
-                },
-            };
-            error!("{}", error_msg);
-            anyhow::bail!(error_msg);
-        }
-
-        let response: message::NewResponse = http_response.json().await?;
-
-        debug!("got: user vm ID={}, gw socket={}", response.user_vm_id, response.gateway_sockaddr);
-
-        // TODO: we need to connect the SocketStream after creating the user VM (and thus adding to
-        // the cold-start time) because currently nanvixd determines the gateway address at
-        // deployment time.
-        let gateway_socktype: SocketType = if *linuxd_deployment == LinuxdDeployment::L2Vm {
-            SocketType::Tcp
-        } else {
-            SocketType::Unix
-        };
-        let gateway_stream: SocketStream = {
-            let deadline: Duration = Duration::from_secs(GATEWAY_CONNECT_TIMEOUT_SECS);
-            match timeout(deadline, async {
-                loop {
-                    let unbound_socket: UnboundSocket = UnboundSocket::new(gateway_socktype);
-                    match unbound_socket.connect(&response.gateway_sockaddr).await {
-                        Ok(stream) => break stream,
-                        Err(_) => {
-                            sleep(Duration::from_millis(GATEWAY_CONNECT_RETRY_SLEEP_MS)).await;
-                            continue;
-                        },
-                    };
-                }
-            })
-            .await
-            {
-                Ok(stream) => stream,
-                Err(_) => {
-                    error!(
-                        "gateway connection to {} timed out after {}s",
-                        response.gateway_sockaddr, GATEWAY_CONNECT_TIMEOUT_SECS
-                    );
-                    anyhow::bail!(
-                        "gateway connection to {} timed out after {}s",
-                        response.gateway_sockaddr,
-                        GATEWAY_CONNECT_TIMEOUT_SECS
-                    );
-                },
-            }
-        };
-        debug!("connected to gateway socket stream");
-
-        Ok((response.user_vm_id, gateway_stream))
-    }
-
-    /// Kill the Nano VM via POST request to nanvixd.
-    pub async fn kill(&mut self, user_vm_id: UserVmIdentifier) -> Result<()> {
-        let mut kill_msg_headers = HeaderMap::new();
-        kill_msg_headers.insert(CONTENT_TYPE, "application/json".parse()?);
-        kill_msg_headers
-            .insert(HTTP_HEADER_MESSAGE_TYPE, format!("{}", MessageType::Kill).parse()?);
-
-        let kill_msg: Kill = Kill { user_vm_id };
-        let response: KillResponse = self
-            .nanvixd_client
-            .post(format!("http://{}", NANVIXD_ADDRESS))
-            .headers(kill_msg_headers)
-            .json(&kill_msg)
-            .send()
-            .await?
-            .json()
-            .await?;
-
-        if response.exit_code != 0 {
-            error!("error killing user VM (id={user_vm_id}, exit-code={})", response.exit_code);
-        }
-
-        Ok(())
-    }
-
-    /// Kill the different components in order.
-    pub fn cleanup(&mut self) {
-        if let Some(nanvixd) = self.nanvixd.as_mut() {
-            debug!("Sending SIGINT to nanvixd");
-            let ret_code: i32 = unsafe { libc::kill(nanvixd.id() as libc::pid_t, libc::SIGINT) };
-
-            if ret_code < 0 {
-                error!("error sending SIGINT to nanvixd: {}", std::io::Error::last_os_error());
-            }
-
-            // Wait for nanvixd to exit with a bounded timeout to prevent indefinite hangs.
-            let deadline = Instant::now() + Duration::from_secs(NANVIXD_SHUTDOWN_TIMEOUT_SECS);
-            loop {
-                match nanvixd.try_wait() {
-                    Ok(Some(exit_status)) => {
-                        if !exit_status.success() {
-                            error!(
-                                "nanvixd returned with non-zero exit status: {:?}",
-                                exit_status.code()
-                            );
-                        }
-                        break;
-                    },
-                    Ok(None) => {
-                        if Instant::now() >= deadline {
-                            warn!(
-                                "nanvixd did not exit within {}s after SIGINT, sending SIGKILL",
-                                NANVIXD_SHUTDOWN_TIMEOUT_SECS
-                            );
-                            let _ =
-                                unsafe { libc::kill(nanvixd.id() as libc::pid_t, libc::SIGKILL) };
-                            let _ = nanvixd.wait();
-                            break;
-                        }
-                        std::thread::sleep(Duration::from_millis(100));
-                    },
-                    Err(e) => {
-                        error!("error waiting for nanvixd: {e:?}");
-                        break;
-                    },
-                }
-            }
-
-            self.nanvixd = None;
-        }
-    }
-
-    ///
-    /// # Description
-    ///
-    /// This method spawns a user VM with the request deployment characteristics, sends an echo
-    /// message and, optionally, records the time it all took, and persists the user VM's stream.
-    ///
-    async fn run_user_vm_echo_once(
-        &mut self,
-        new_msg_headers: HeaderMap,
-        new_msg: New,
-        linuxd_deployment: &LinuxdDeployment,
-        cleanup_duration: Duration,
-        latencies: Option<&mut Vec<u128>>,
-        in_flight_uvms: &mut Option<Vec<(UserVmIdentifier, SocketStream)>>,
-    ) -> Result<()> {
-        let payload: [u8; DEFAULT_PAYLOAD_SIZE] = [7u8; DEFAULT_PAYLOAD_SIZE];
-        let mut response_payload: [u8; DEFAULT_PAYLOAD_SIZE] = [0u8; DEFAULT_PAYLOAD_SIZE];
-
-        let start: Instant = Instant::now();
-        let (user_vm_id, mut gateway_stream): (UserVmIdentifier, SocketStream) = self
-            .start(new_msg, new_msg_headers, linuxd_deployment)
-            .await?;
-        let echo_io_timeout: Duration = Duration::from_secs(ECHO_IO_TIMEOUT_SECS);
-        timeout(echo_io_timeout, gateway_stream.write_all(&payload))
-            .await
-            .map_err(|_| {
-                error!("echo write timed out after {}s", ECHO_IO_TIMEOUT_SECS);
-                anyhow::anyhow!("echo write timed out after {}s", ECHO_IO_TIMEOUT_SECS)
-            })?
-            .map_err(|e| anyhow::anyhow!("echo write failed: {e}"))?;
-        timeout(echo_io_timeout, gateway_stream.read_exact(&mut response_payload))
-            .await
-            .map_err(|_| {
-                error!("echo read timed out after {}s", ECHO_IO_TIMEOUT_SECS);
-                anyhow::anyhow!("echo read timed out after {}s", ECHO_IO_TIMEOUT_SECS)
-            })?
-            .map_err(|e| anyhow::anyhow!("echo read failed: {e}"))?;
-        let elapsed_micros: u128 = start.elapsed().as_micros();
-
-        // Only record latency if requested to.
-        if let Some(latencies) = latencies {
-            latencies.push(elapsed_micros);
-        }
-
-        // Sanity-check the message to make sure is the same we sent.
-        if response_payload != payload {
-            error!("received payload does not match sent payload!");
-            error!(" - sent: {payload:?}");
-            error!(" - got: {response_payload:?}");
-        }
-
-        // If we must persist, store the user VM id and gateway stream. Otherwise clean-up.
-        if let Some(in_flight_uvms) = in_flight_uvms.as_mut() {
-            in_flight_uvms.push((user_vm_id, gateway_stream));
-        } else {
-            // Kill the user VM.
-            self.kill(user_vm_id).await?;
-        }
-
-        sleep(cleanup_duration).await;
-
-        Ok(())
-    }
-
-    //==============================================================================================
-    // Main entrypoint functions for each benchmark
-    //==============================================================================================
-
-    /// This function runs the boot-time experiment, where we measure the time to start a user VM
-    /// with a noop application and exit. To properly isolate just the time to start a user VM, we
-    /// do not make use of nanvixd here. Instead, we start the user VM manually.
-    pub async fn run_boot_time(&mut self) -> Result<()> {
-        // Display a progress bar
-        let pb = ProgressBar::new(self.iterations.try_into().unwrap());
-        pb.set_style(
-            ProgressStyle::default_bar()
-                .template("{msg} [{bar:40.cyan/blue}] {pos}/{len} ({percent}%)")
-                .expect("error creating progress bar")
-                .progress_chars("#>-"),
-        );
-        pb.set_message("Benchmark progress:");
-
-        let mut latencies: Vec<u128> = Vec::with_capacity(self.iterations);
-        for _ in 0..self.iterations {
-            let (vcpu_thread_stdout_tx, mut vcpu_thread_stdout_rx) =
-                mpsc::channel::<IkcFrame>(CHANNEL_CAPACITY);
-            let stdout_drain: JoinHandle<()> =
-                ::tokio::spawn(
-                    async move { while vcpu_thread_stdout_rx.recv().await.is_some() {} },
-                );
-
-            let (io_control_command_tx, io_control_rx) =
-                mpsc::channel::<IoControlCommand>(CHANNEL_CAPACITY);
-            let (io_control_tx, mut io_control_response_rx) =
-                mpsc::channel::<IoControlResponse>(CHANNEL_CAPACITY);
-            let io_response_drain: JoinHandle<()> =
-                ::tokio::spawn(
-                    async move { while io_control_response_rx.recv().await.is_some() {} },
-                );
-
-            let (io_thread_data_tx, memory_thread_data_rx) =
-                mpsc::channel::<IkcFrame>(CHANNEL_CAPACITY);
-
-            let kernel_filename: String =
-                format!("{}/bin/kernel.elf", self.workspace_root.display());
-            let initrd_filename: String = self.flavour.get_program(&self.workspace_root);
-
-            // Create shared counters for tracking message flow across threads.
-            let counters: MessageCounters = MessageCounters::new();
-
-            let start = Instant::now();
-            let user_vm_handle = UserVm::spawn(UserVmArgs {
-                kernel_filename,
-                initrd_filename: Some(initrd_filename),
-                initrd_args: None,
-                ramfs_filename: None,
-                stderr: Some("/dev/null".to_string()),
-                vcpu_thread_stdout_tx,
-                memory_thread_data_rx,
-                io_control_rx,
-                io_control_tx,
-                counters,
-                snapshot_path: None,
-                #[cfg(feature = "gdb")]
-                gdb_port: None,
-            });
-
-            let join_result = user_vm_handle.await;
-
-            drop(io_thread_data_tx);
-            drop(io_control_command_tx);
-
-            if let Err(error) = stdout_drain.await {
-                error!("error draining user VM stdout channel: {error:?}");
-            }
-            if let Err(error) = io_response_drain.await {
-                error!("error draining user VM control channel: {error:?}");
-            }
-
-            match join_result {
-                Ok(Ok(exit_status)) => {
-                    if exit_status != 0 {
-                        let reason: String =
-                            format!("error running user VM, exit-status={exit_status}");
-                        error!("{reason}");
-                        return Err(anyhow::anyhow!(reason));
-                    }
-                    debug!("User VM: done running");
-                },
-                Ok(Err(error)) => {
-                    error!("error running user VM: {error:?}");
-                    return Err(error);
-                },
-                Err(error) => {
-                    let reason: String = format!("error joining user VM task: {error:?}");
-                    error!("{reason}");
-                    return Err(anyhow::anyhow!(reason));
-                },
-            }
-
-            latencies.push(start.elapsed().as_micros());
-
-            pb.inc(1);
-
-            // Need to give some time to clean-up
-            sleep(Duration::from_millis(CLEANUP_SLEEP_DURATION)).await;
-        }
-
-        pb.finish();
-        latencies.sort();
-        println!("p50: {} us", latencies[(self.iterations * 50) / 100]);
-        println!("p95: {} us", latencies[(self.iterations * 95) / 100]);
-        println!("p99: {} us", latencies[(self.iterations * 99) / 100]);
-
-        Ok(())
-    }
-
-    ///
-    /// # Description
-    ///
-    /// Runs the snapshot-restore experiment. The setup phase creates a snapshot using the
-    /// `snapshot-rust-nostd.elf` program, and the measurement phase iteratively restores from
-    /// that snapshot, measuring the time for each restore-and-exit cycle.
-    ///
-    pub async fn run_snapshot_restore(&mut self) -> Result<()> {
-        let kernel_filename: String = format!("{}/bin/kernel.elf", self.workspace_root.display());
-        let snapshot_program: String = self.flavour.get_program(&self.workspace_root);
-
-        // Ensure snapshots directory exists.
-        let snapshots_dir: std::path::PathBuf = self.workspace_root.join("snapshots");
-        if !snapshots_dir.exists() {
-            std::fs::create_dir_all(&snapshots_dir)?;
-        }
-
-        // Phase 1: Create a snapshot by running the snapshot program.
-        println!("Creating snapshot...");
-        {
-            let (vcpu_thread_stdout_tx, mut vcpu_thread_stdout_rx) =
-                mpsc::channel::<IkcFrame>(CHANNEL_CAPACITY);
-            let stdout_drain: JoinHandle<()> =
-                ::tokio::spawn(
-                    async move { while vcpu_thread_stdout_rx.recv().await.is_some() {} },
-                );
-
-            let (io_control_command_tx, io_control_rx) =
-                mpsc::channel::<IoControlCommand>(CHANNEL_CAPACITY);
-            let (io_control_tx, mut io_control_response_rx) =
-                mpsc::channel::<IoControlResponse>(CHANNEL_CAPACITY);
-            let io_response_drain: JoinHandle<()> =
-                ::tokio::spawn(
-                    async move { while io_control_response_rx.recv().await.is_some() {} },
-                );
-
-            let (_io_thread_data_tx, memory_thread_data_rx) =
-                mpsc::channel::<IkcFrame>(CHANNEL_CAPACITY);
-
-            let counters: MessageCounters = MessageCounters::new();
-
-            let user_vm_handle = UserVm::spawn(UserVmArgs {
-                kernel_filename: kernel_filename.clone(),
-                initrd_filename: Some(snapshot_program),
-                initrd_args: None,
-                ramfs_filename: None,
-                stderr: Some("/dev/null".to_string()),
-                vcpu_thread_stdout_tx,
-                memory_thread_data_rx,
-                io_control_rx,
-                io_control_tx,
-                counters,
-                snapshot_path: None,
-                #[cfg(feature = "gdb")]
-                gdb_port: None,
-            });
-
-            let join_result = user_vm_handle.await;
-
-            drop(io_control_command_tx);
-
-            if let Err(error) = stdout_drain.await {
-                error!("error draining user VM stdout channel: {error:?}");
-            }
-            if let Err(error) = io_response_drain.await {
-                error!("error draining user VM control channel: {error:?}");
-            }
-
-            match join_result {
-                Ok(Ok(exit_status)) => {
-                    if exit_status != 0 {
-                        let reason: String =
-                            format!("error creating snapshot, exit-status={exit_status}");
-                        error!("{reason}");
-                        return Err(anyhow::anyhow!(reason));
-                    }
-                    debug!("Snapshot created successfully");
-                },
-                Ok(Err(error)) => {
-                    error!("error creating snapshot: {error:?}");
-                    return Err(error);
-                },
-                Err(error) => {
-                    let reason: String = format!("error joining snapshot VM task: {error:?}");
-                    error!("{reason}");
-                    return Err(anyhow::anyhow!(reason));
-                },
-            }
-
-            sleep(Duration::from_millis(CLEANUP_SLEEP_DURATION)).await;
-        }
-
-        // Phase 2: Measure snapshot restore latency.
-        let pb = ProgressBar::new(self.iterations.try_into().unwrap());
-        pb.set_style(
-            ProgressStyle::default_bar()
-                .template("{msg} [{bar:40.cyan/blue}] {pos}/{len} ({percent}%)")
-                .expect("error creating progress bar")
-                .progress_chars("#>-"),
-        );
-        pb.set_message("Snapshot restore:");
-
-        let mut latencies: Vec<u128> = Vec::with_capacity(self.iterations);
-        for _ in 0..self.iterations {
-            let (vcpu_thread_stdout_tx, mut vcpu_thread_stdout_rx) =
-                mpsc::channel::<IkcFrame>(CHANNEL_CAPACITY);
-            let stdout_drain: JoinHandle<()> =
-                ::tokio::spawn(
-                    async move { while vcpu_thread_stdout_rx.recv().await.is_some() {} },
-                );
-
-            let (io_control_command_tx, io_control_rx) =
-                mpsc::channel::<IoControlCommand>(CHANNEL_CAPACITY);
-            let (io_control_tx, mut io_control_response_rx) =
-                mpsc::channel::<IoControlResponse>(CHANNEL_CAPACITY);
-            let io_response_drain: JoinHandle<()> =
-                ::tokio::spawn(
-                    async move { while io_control_response_rx.recv().await.is_some() {} },
-                );
-
-            let (_io_thread_data_tx, memory_thread_data_rx) =
-                mpsc::channel::<IkcFrame>(CHANNEL_CAPACITY);
-
-            let counters: MessageCounters = MessageCounters::new();
-
-            let start = Instant::now();
-            let user_vm_handle = UserVm::spawn(UserVmArgs {
-                kernel_filename: kernel_filename.clone(),
-                initrd_filename: None,
-                initrd_args: None,
-                ramfs_filename: None,
-                stderr: Some("/dev/null".to_string()),
-                vcpu_thread_stdout_tx,
-                memory_thread_data_rx,
-                io_control_rx,
-                io_control_tx,
-                counters,
-                snapshot_path: Some(kernel_filename.clone()),
-                #[cfg(feature = "gdb")]
-                gdb_port: None,
-            });
-
-            let join_result = user_vm_handle.await;
-
-            drop(io_control_command_tx);
-
-            if let Err(error) = stdout_drain.await {
-                error!("error draining user VM stdout channel: {error:?}");
-            }
-            if let Err(error) = io_response_drain.await {
-                error!("error draining user VM control channel: {error:?}");
-            }
-
-            match join_result {
-                Ok(Ok(exit_status)) => {
-                    if exit_status != 0 {
-                        let reason: String =
-                            format!("error restoring snapshot, exit-status={exit_status}");
-                        error!("{reason}");
-                        return Err(anyhow::anyhow!(reason));
-                    }
-                    debug!("Snapshot restore: done running");
-                },
-                Ok(Err(error)) => {
-                    error!("error restoring snapshot: {error:?}");
-                    return Err(error);
-                },
-                Err(error) => {
-                    let reason: String = format!("error joining snapshot restore task: {error:?}");
-                    error!("{reason}");
-                    return Err(anyhow::anyhow!(reason));
-                },
-            }
-
-            latencies.push(start.elapsed().as_micros());
-
-            pb.inc(1);
-
-            sleep(Duration::from_millis(CLEANUP_SLEEP_DURATION)).await;
-        }
-
-        pb.finish();
-        latencies.sort();
-        println!("p50: {} us", latencies[(self.iterations * 50) / 100]);
-        println!("p95: {} us", latencies[(self.iterations * 95) / 100]);
-        println!("p99: {} us", latencies[(self.iterations * 99) / 100]);
-
-        Ok(())
-    }
-
-    ///
-    /// # Description
-    ///
-    /// This function runs the cold-start experiment, where we measure the time to start linuxd,
-    /// start a VM, and send a request to the new VM.
-    ///
-    /// # Arguments
-    ///
-    /// - `linuxd_deployment`: deployment mode for linuxd.
-    /// - `uservm_deployment`: deployment mode for the user VM.
-    ///
-    pub async fn run_cold_start(
-        &mut self,
-        linuxd_deployment: &LinuxdDeployment,
-        user_vm_deployment: &UserVmDeployment,
-    ) -> Result<()> {
-        // Start nanvixd once.
-        self.setup(linuxd_deployment);
-
-        // Display a progress bar
-        let pb: ProgressBar = ProgressBar::new(self.iterations.try_into().unwrap());
-        pb.set_style(
-            ProgressStyle::default_bar()
-                .template("{msg} [{bar:40.cyan/blue}] {pos}/{len} ({percent}%)")
-                .expect("error creating progress bar")
-                .progress_chars("#>-"),
-        );
-        pb.set_message("Benchmark progress:");
-
-        // Work-out the cleanup sleep duration depending on the linuxd deployment mode.
-        let cleanup_sleep_duration: Duration = if *linuxd_deployment == LinuxdDeployment::L2Vm {
-            Duration::from_millis(CLEANUP_L2_SLEEP_DURATION)
-        } else {
-            Duration::from_millis(CLEANUP_SLEEP_DURATION)
-        };
-
-        let mut latencies: Vec<u128> = Vec::with_capacity(self.iterations);
-        for iter in 0..self.iterations {
-            // In this benchmark we measure the time to start both linuxd and the user VM, so we
-            // give each iteration a different tenant id and application name.
-            let tenant_id = format!("{DEFAULT_TENANT_ID}-{iter}");
-            let app_name = format!("{DEFAULT_APP_NAME}-{iter}");
-
-            // Get the right message for this new user VM.
-            let (new_msg_headers, new_msg): (HeaderMap, New) =
-                self.prepare_new_message(Some(tenant_id), Some(app_name))?;
-
-            // In case we are pre-warming, we will run the user VM once without keeping track of
-            // the time-elapsed.
-            if *user_vm_deployment == UserVmDeployment::PreWarm {
-                self.run_user_vm_echo_once(
-                    new_msg_headers.clone(),
-                    new_msg.clone(),
-                    linuxd_deployment,
-                    cleanup_sleep_duration,
-                    None,
-                    &mut None,
-                )
-                .await?;
-            }
-
-            self.run_user_vm_echo_once(
-                new_msg_headers.clone(),
-                new_msg.clone(),
-                linuxd_deployment,
-                cleanup_sleep_duration,
-                Some(&mut latencies),
-                &mut None,
-            )
-            .await?;
-            pb.inc(1);
-        }
-
-        pb.finish();
-        println!("First req: {} us", latencies[0]);
-        latencies.sort();
-        println!("p50: {} us", latencies[(self.iterations as f32 * 0.5) as usize]);
-        println!("p95: {} us", latencies[(self.iterations as f32 * 0.95) as usize]);
-        println!("p99: {} us", latencies[(self.iterations as f32 * 0.99) as usize]);
-
-        print!("Cleaning up...");
-        self.cleanup();
-        println!("done!");
-
-        Ok(())
-    }
-
-    ///
-    /// # Description
-    ///
-    /// This benchmark measures the time to start N concurrent user VMs, all sharing the same
-    /// linuxd instance.
-    ///
-    /// # Arguments
-    ///
-    /// - `linuxd_deployment`: deployment mode for linuxd.
-    /// - `num_concurrent_vms`: number of VMs to start concurrently.
-    ///
-    pub async fn run_concurrent(
-        &mut self,
-        linuxd_deployment: &LinuxdDeployment,
-        num_concurrent_vms: usize,
-    ) -> Result<()> {
-        // Display a progress bar
-        let pb: ProgressBar = ProgressBar::new(num_concurrent_vms.try_into().unwrap());
-        pb.set_style(
-            ProgressStyle::default_bar()
-                .template("{msg} [{bar:40.cyan/blue}] {pos}/{len} ({percent}%)")
-                .expect("error creating progress bar")
-                .progress_chars("#>-"),
-        );
-        pb.set_message("Benchmark progress:");
-
-        // Start nanvixd once.
-        self.setup(linuxd_deployment);
-
-        // Work-out the cleanup sleep duration depending on the linuxd deployment mode.
-        let cleanup_sleep_duration: Duration = if *linuxd_deployment == LinuxdDeployment::L2Vm {
-            Duration::from_millis(CLEANUP_L2_SLEEP_DURATION)
-        } else {
-            Duration::from_millis(CLEANUP_SLEEP_DURATION)
-        };
-
-        let mut latencies: Vec<u128> = Vec::with_capacity(num_concurrent_vms);
-        let mut in_flight_uvms: Option<Vec<(UserVmIdentifier, SocketStream)>> =
-            Some(Vec::with_capacity(num_concurrent_vms));
-        for iter in 0..num_concurrent_vms {
-            // In this benchmark we want all user VMs to share the same linuxd instance, so they
-            // run concurrently. We therefore keep the tenant id constant (default) and give
-            // each user VM a different name.
-            let app_name: String = format!("bar-{iter}");
-            let (new_msg_headers, new_msg): (HeaderMap, New) =
-                self.prepare_new_message(None, Some(app_name))?;
-
-            // We want all user VMs to run concurrently, so we pass an in-flight map to keep them
-            // around instead of killing them after getting the echo.
-            self.run_user_vm_echo_once(
-                new_msg_headers,
-                new_msg,
-                linuxd_deployment,
-                cleanup_sleep_duration,
-                Some(&mut latencies),
-                &mut in_flight_uvms,
-            )
-            .await?;
-            pb.inc(1);
-        }
-
-        pb.finish();
-        println!(
-            "Time to spawn {num_concurrent_vms} user VMs (in serial): {:.2} s",
-            (latencies.iter().sum::<u128>() as f64) / 1_000_000.0
-        );
-        latencies.sort();
-        println!("p50: {} us", latencies[(num_concurrent_vms as f32 * 0.5) as usize]);
-        println!("p95: {} us", latencies[(num_concurrent_vms as f32 * 0.95) as usize]);
-        println!("p99: {} us", latencies[(num_concurrent_vms as f32 * 0.99) as usize]);
-
-        print!("Cleaning up...");
-
-        if let Some(in_flight_uvms) = in_flight_uvms.as_mut() {
-            for (user_vm_id, _) in in_flight_uvms.drain(..) {
-                self.kill(user_vm_id).await?;
-            }
-        } else {
-            error!("in_flight_uvms cannot be none");
-        }
-        self.cleanup();
-
-        println!("done!");
-
-        Ok(())
-    }
-
-    ///
-    /// # Description
-    ///
-    /// This function runs the round-trip latency benchmark, where we measure the latency of
-    /// sending one message and getting it back, as we increase the message size.
-    ///
-    /// Given that we have many message sizes (hence many rows in the result table) we default to
-    /// reporting only one percentile.
-    ///
-    /// # Arguments
-    ///
-    /// - `linuxd_deployment`: deployment mode for linuxd.
-    /// - `percentile`: what percentile to report.
-    ///
-    pub async fn run_round_trip_latency(
-        &mut self,
-        linuxd_deployment: &LinuxdDeployment,
-    ) -> Result<()> {
-        let message_sizes: Vec<(&str, u64)> = vec![
-            ("32 B", 32),
-            ("64 B", 64),
-            ("128 B", 128),
-            ("256 B", 256),
-            ("512 B", 512),
-            ("1 KiB", 1024),
-            ("4 KiB", 4 * 1024),
-        ];
-
-        // Display a progress bar
-        let total_num_iters: usize = self.iterations * message_sizes.len();
-        let pb: ProgressBar = ProgressBar::new(total_num_iters as u64);
-        pb.set_style(
-            ProgressStyle::default_bar()
-                .template("{msg} [{bar:40.cyan/blue}] {pos}/{len} ({percent}%)")
-                .expect("error creating progress bar")
-                .progress_chars("#>-"),
-        );
-        pb.set_message("Benchmark progress:");
-
-        let (new_msg_headers, new_msg): (HeaderMap, New) = self.prepare_new_message(None, None)?;
-
-        // Start nanvixd.
-        self.setup(linuxd_deployment);
-
-        let mut latencies: HashMap<&str, Vec<u128>> = HashMap::new();
-        let user_vm_id = {
-            // Start User VM.
-            let (user_vm_id, mut gateway_stream) = self
-                .start(new_msg, new_msg_headers, linuxd_deployment)
-                .await?;
-
-            // Iterate over all possible message sizes.
-            for (label, message_size) in &message_sizes {
-                let payload: Vec<u8> = vec![7u8; *message_size as usize];
-
-                // For each message size send many messages to get statistically relevant results.
-                for _ in 0..self.iterations {
-                    let mut response_payload: Vec<u8> = vec![0u8; *message_size as usize];
-
-                    let start: Instant = Instant::now();
-                    gateway_stream.write_all(&payload).await?;
-                    gateway_stream.read_exact(&mut response_payload).await?;
-                    latencies
-                        .entry(label)
-                        .or_default()
-                        .push(start.elapsed().as_micros());
-
-                    // Sanity-check the message to make sure is the same we sent.
-                    if response_payload != payload {
-                        error!("received payload does not match sent payload!");
-                        error!(" - sent: {payload:?}");
-                        error!(" - got: {response_payload:?}");
-                    }
-
-                    pb.inc(1);
-                    sleep(Duration::from_millis(CLEANUP_SLEEP_DURATION)).await;
-                }
-            }
-            user_vm_id
-        };
-
-        // Kill the user VM.
-        self.kill(user_vm_id).await?;
-
-        // Stop nanvixd.
-        self.cleanup();
-
-        pb.finish();
-        println!("Size:\tp50\tp95\tp99 [us]");
-        // Iterate over the message size list to print the labels in order.
-        for (label, _) in message_sizes.iter() {
-            if let Some(latencies) = latencies.get_mut(label) {
-                latencies.sort();
-                let p50: u128 = latencies[(self.iterations as f32 * 0.5) as usize];
-                let p95: u128 = latencies[(self.iterations as f32 * 0.95) as usize];
-                let p99: u128 = latencies[(self.iterations as f32 * 0.99) as usize];
-                println!("{label}:\t{p50}\t{p95}\t{p99}");
-            } else {
-                warn!("missing latencies for message size: {label}");
-            }
-        }
-
-        Ok(())
-    }
-
-    /// This function runs the warm start benchmark, where we measure the time to send a request
-    /// into the VM once it has started executing.
-    pub async fn run_warm_start(&mut self, linuxd_deployment: &LinuxdDeployment) -> Result<()> {
-        // Display a progress bar
-        let pb = ProgressBar::new(self.iterations.try_into().unwrap());
-        pb.set_style(
-            ProgressStyle::default_bar()
-                .template("{msg} [{bar:40.cyan/blue}] {pos}/{len} ({percent}%)")
-                .expect("error creating progress bar")
-                .progress_chars("#>-"),
-        );
-        pb.set_message("Benchmark progress:");
-
-        // Payload we are sending over the wire
-        let payload = [7u8; DEFAULT_PAYLOAD_SIZE];
-
-        let (new_msg_headers, new_msg) = self.prepare_new_message(None, None)?;
-
-        // Start nanvixd.
-        self.setup(linuxd_deployment);
-
-        // Start User VM.
-        let mut latencies: Vec<u128> = Vec::with_capacity(self.iterations);
-        let user_vm_id = {
-            let (user_vm_id, mut gateway_stream) = self
-                .start(new_msg, new_msg_headers, linuxd_deployment)
-                .await?;
-
-            // Warmup: send one untimed echo to trigger lazy initialization (worker thread
-            // creation, TCP path warm-up, etc.) so that timed iterations reflect steady-state
-            // latency.
-            {
-                let mut warmup_response = [0u8; DEFAULT_PAYLOAD_SIZE];
-                gateway_stream.write_all(&payload).await?;
-                gateway_stream.read_exact(&mut warmup_response).await?;
-                sleep(Duration::from_millis(WARMUP_SLEEP_DURATION)).await;
-            }
-
-            for _ in 0..self.iterations {
-                let mut response_payload = [0u8; DEFAULT_PAYLOAD_SIZE];
-
-                let start = Instant::now();
-                gateway_stream.write_all(&payload).await?;
-                gateway_stream.read_exact(&mut response_payload).await?;
-                latencies.push(start.elapsed().as_micros());
-
-                // Sanity-check the message to make sure is the same we sent.
-                if response_payload != payload {
-                    error!("received payload does not match sent payload!");
-                    error!(" - sent: {payload:?}");
-                    error!(" - got: {response_payload:?}");
-                }
-
-                pb.inc(1);
-                sleep(Duration::from_millis(CLEANUP_SLEEP_DURATION)).await;
-            }
-            user_vm_id
-        };
-
-        // Kill the user VM.
-        self.kill(user_vm_id).await?;
-
-        // Stop nanvixd.
-        self.cleanup();
-
-        pb.finish();
-        println!("First req: {} us", latencies[0]);
-        latencies.sort();
-        println!("p50: {} us", latencies[(self.iterations as f32 * 0.5) as usize]);
-        println!("p95: {} us", latencies[(self.iterations as f32 * 0.95) as usize]);
-        println!("p99: {} us", latencies[(self.iterations as f32 * 0.99) as usize]);
-
-        Ok(())
-    }
-
-    /// In this micro-benchmark we measure the time for a message to travel
-    /// all the way from the VMM to the guest application and back. To achieve
-    /// this, we connect the user VM to a gateway that emulates linuxd.
-    pub async fn run_warm_start_vmm(&mut self) -> Result<()> {
-        // Display a progress bar.
-        let pb = ProgressBar::new(self.iterations.try_into().unwrap());
-        pb.set_style(
-            ProgressStyle::default_bar()
-                .template("{msg} [{bar:40.cyan/blue}] {pos}/{len} ({percent}%)")
-                .expect("error creating progress bar")
-                .progress_chars("#>-"),
-        );
-        pb.set_message("Benchmark progress:");
-
-        // Payload we are sending over the wire.
-        let data = [7u8; DEFAULT_PAYLOAD_SIZE];
-        let mut payload: Vec<u8> = Vec::with_capacity(mem::size_of::<u32>() + data.len());
-        payload.extend_from_slice(&DEFAULT_PAYLOAD_SIZE.to_le_bytes());
-        payload.extend_from_slice(&data);
-        let (vcpu_thread_stdout_tx, mut vcpu_thread_stdout_rx) =
-            mpsc::channel::<IkcFrame>(CHANNEL_CAPACITY);
-        let (io_thread_data_tx, memory_thread_data_rx) =
-            mpsc::channel::<IkcFrame>(CHANNEL_CAPACITY);
-        let (io_control_command_tx, io_control_rx) =
-            mpsc::channel::<IoControlCommand>(CHANNEL_CAPACITY);
-        let (io_control_tx, mut io_control_response_rx) =
-            mpsc::channel::<IoControlResponse>(CHANNEL_CAPACITY);
-
-        let kernel_filename: String = format!("{}/bin/kernel.elf", self.workspace_root.display());
-        let program: String = self.flavour.get_program(&self.workspace_root);
-
-        // Create shared counters for tracking message flow across threads.
-        let counters: MessageCounters = MessageCounters::new();
-
-        let user_vm_handle = UserVm::spawn(UserVmArgs {
-            kernel_filename,
-            initrd_filename: Some(program),
-            initrd_args: None,
-            ramfs_filename: None,
-            stderr: Some("/dev/null".to_string()),
-            vcpu_thread_stdout_tx,
-            memory_thread_data_rx,
-            io_control_rx,
-            io_control_tx,
-            counters,
-            snapshot_path: None,
-            #[cfg(feature = "gdb")]
-            gdb_port: None,
-        });
-
-        // Warmup: run one untimed echo cycle through the full IKC protocol to trigger
-        // lazy initialization so that timed iterations reflect steady-state latency.
-        {
-            // Step 1: Receive the ReadRequest IKC message from the guest.
-            let warmup_read_msg: Message = match vcpu_thread_stdout_rx.recv().await {
-                Some(IkcFrame::Message(message)) => message,
-                Some(IkcFrame::Bulk(_)) => {
-                    anyhow::bail!("warmup: unexpected bulk during ReadRequest")
-                },
-                None => anyhow::bail!("warmup: channel closed during ReadRequest"),
-            };
-            let warmup_linuxd_msg: LinuxDaemonMessage =
-                LinuxDaemonMessage::try_from_bytes(warmup_read_msg.payload)
-                    .map_err(|_| anyhow::anyhow!("warmup: error parsing LinuxDaemon message"))?;
-            // NOTE: `as_id()` returns `Err(ThreadIdentifier)` when the source is a
-            // thread (expected) and `Ok(ProcessIdentifier)` when it is a process.
-            let warmup_tid: ThreadIdentifier = match { warmup_read_msg.source }.as_id() {
-                Err(tid) => tid,
-                Ok(pid) => anyhow::bail!("warmup: unexpected message source: {pid:?}"),
-            };
-            let _warmup_read_req: ReadRequest = ReadRequest::from_bytes(warmup_linuxd_msg.payload);
-
-            // Step 2: Receive the bulk pull request from the guest kernel.
-            let warmup_pull_header: DataChunkHeader = match vcpu_thread_stdout_rx.recv().await {
-                Some(IkcFrame::Bulk(bulk)) => *bulk.header(),
-                Some(IkcFrame::Message(_)) => {
-                    anyhow::bail!("warmup: unexpected message during bulk pull")
-                },
-                None => anyhow::bail!("warmup: channel closed during bulk pull"),
-            };
-
-            // Step 3: Send the bulk data response back to the kernel buffer.
-            let warmup_bulk_response: DataChunk = DataChunk::new(
-                DataChunkHeader::new(
-                    warmup_pull_header.source_pid(),
-                    warmup_pull_header.source_tid(),
-                    warmup_pull_header.destination_pid(),
-                    warmup_pull_header.destination_tid(),
-                    warmup_pull_header.data_addr(),
-                    payload.len() as u32,
-                ),
-                payload.clone(),
-            );
-            io_thread_data_tx
-                .send(IkcFrame::Bulk(warmup_bulk_response))
-                .await?;
-
-            // Step 4: Send ReadResponse metadata.
-            let warmup_empty_buf: [u8; ReadResponse::BUFFER_SIZE] =
-                [0u8; ReadResponse::BUFFER_SIZE];
-            let warmup_read_response: Message =
-                ReadResponse::build(warmup_tid, payload.len() as i32, warmup_empty_buf);
-            io_thread_data_tx
-                .send(IkcFrame::Message(warmup_read_response))
-                .await?;
-
-            // Step 5: Receive the WriteRequest IKC message from the guest.
-            match vcpu_thread_stdout_rx.recv().await {
-                Some(IkcFrame::Message(_)) => {},
-                Some(IkcFrame::Bulk(_)) => {
-                    anyhow::bail!("warmup: unexpected bulk during WriteRequest")
-                },
-                None => anyhow::bail!("warmup: channel closed during WriteRequest"),
-            };
-
-            // Step 6: Receive the bulk push data from the guest.
-            match vcpu_thread_stdout_rx.recv().await {
-                Some(IkcFrame::Bulk(_)) => {},
-                Some(IkcFrame::Message(_)) => {
-                    anyhow::bail!("warmup: unexpected message during bulk push")
-                },
-                None => anyhow::bail!("warmup: channel closed during bulk push"),
-            };
-
-            // Step 7: Send WriteResponse to acknowledge the write.
-            let warmup_write_response: Message =
-                WriteResponse::build(warmup_tid, payload.len() as i32);
-            io_thread_data_tx
-                .send(IkcFrame::Message(warmup_write_response))
-                .await?;
-
-            sleep(Duration::from_millis(WARMUP_SLEEP_DURATION)).await;
-        }
-
-        let mut latencies: Vec<u128> = Vec::with_capacity(self.iterations);
-        for _ in 0..self.iterations {
-            // Step 1: Receive the ReadRequest IKC message from the guest.
-            let ipc_read_message: Message = match vcpu_thread_stdout_rx.recv().await {
-                Some(IkcFrame::Message(message)) => message,
-                Some(IkcFrame::Bulk(_)) => {
-                    let reason: String = "unexpected data chunk transfer received while waiting \
-                                          for ReadRequest"
-                        .to_string();
-                    error!("run_warm_start_vmm(): {reason}");
-                    anyhow::bail!(reason);
-                },
-                None => {
-                    let reason: String = "user VM channel closed unexpectedly while waiting for \
-                                          ReadRequest"
-                        .to_string();
-                    error!("run_warm_start_vmm(): {reason}");
-                    anyhow::bail!(reason);
-                },
-            };
-            let linuxd_message: LinuxDaemonMessage =
-                match LinuxDaemonMessage::try_from_bytes(ipc_read_message.payload) {
-                    Ok(message) => message,
-                    Err(_) => {
-                        return Err(anyhow::anyhow!(
-                            "Error parsing IPC message to LinuxDaemon message"
-                        ));
-                    },
-                };
-            let tid: ThreadIdentifier = match { ipc_read_message.source }.as_id() {
-                Err(tid) => tid,
-                Ok(pid) => return Err(anyhow::anyhow!("unexpected message source: {pid:?}")),
-            };
-            let _read_request: ReadRequest = ReadRequest::from_bytes(linuxd_message.payload);
-
-            // Step 2: Receive the bulk pull request from the guest kernel.
-            let pull_header: DataChunkHeader = match vcpu_thread_stdout_rx.recv().await {
-                Some(IkcFrame::Bulk(bulk)) => *bulk.header(),
-                Some(IkcFrame::Message(_)) => {
-                    let reason: String = "unexpected IKC message received while waiting for bulk \
-                                          pull request"
-                        .to_string();
-                    error!("run_warm_start_vmm(): {reason}");
-                    anyhow::bail!(reason);
-                },
-                None => {
-                    let reason: String = "user VM channel closed unexpectedly while waiting for \
-                                          bulk pull request"
-                        .to_string();
-                    error!("run_warm_start_vmm(): {reason}");
-                    anyhow::bail!(reason);
-                },
-            };
-
-            // Now we are ready to push bulk data and ReadResponse, and wait for a WriteRequest.
-            let start = Instant::now();
-
-            // Step 3: Send the bulk data response back to the kernel buffer.
-            let bulk_response: DataChunk = DataChunk::new(
-                DataChunkHeader::new(
-                    pull_header.source_pid(),
-                    pull_header.source_tid(),
-                    pull_header.destination_pid(),
-                    pull_header.destination_tid(),
-                    pull_header.data_addr(),
-                    payload.len() as u32,
-                ),
-                payload.clone(),
-            );
-            io_thread_data_tx
-                .send(IkcFrame::Bulk(bulk_response))
-                .await?;
-
-            // Step 4: Send ReadResponse metadata (buffer is empty; data was sent via bulk).
-            let empty_buf: [u8; ReadResponse::BUFFER_SIZE] = [0u8; ReadResponse::BUFFER_SIZE];
-            let read_response: Message = ReadResponse::build(tid, payload.len() as i32, empty_buf);
-            io_thread_data_tx
-                .send(IkcFrame::Message(read_response))
-                .await?;
-
-            // Step 5: Receive the WriteRequest IKC message from the guest.
-            let _write_request: Message = match vcpu_thread_stdout_rx.recv().await {
-                Some(IkcFrame::Message(message)) => message,
-                Some(IkcFrame::Bulk(_)) => {
-                    let reason: String = "unexpected data chunk transfer received while waiting \
-                                          for WriteRequest"
-                        .to_string();
-                    error!("run_warm_start_vmm(): {reason}");
-                    anyhow::bail!(reason);
-                },
-                None => {
-                    let reason: String = "user VM channel closed unexpectedly while waiting for \
-                                          WriteRequest"
-                        .to_string();
-                    error!("run_warm_start_vmm(): {reason}");
-                    anyhow::bail!(reason);
-                },
-            };
-
-            // Step 6: Receive the bulk push data from the guest.
-            let _push_data: DataChunk = match vcpu_thread_stdout_rx.recv().await {
-                Some(IkcFrame::Bulk(bulk)) => bulk,
-                Some(IkcFrame::Message(_)) => {
-                    let reason: String = "unexpected IKC message received while waiting for bulk \
-                                          push data"
-                        .to_string();
-                    error!("run_warm_start_vmm(): {reason}");
-                    anyhow::bail!(reason);
-                },
-                None => {
-                    let reason: String = "user VM channel closed unexpectedly while waiting for \
-                                          bulk push data"
-                        .to_string();
-                    error!("run_warm_start_vmm(): {reason}");
-                    anyhow::bail!(reason);
-                },
-            };
-
-            latencies.push(start.elapsed().as_micros());
-
-            // Step 7: Send WriteResponse to acknowledge the write.
-            let write_response: Message = WriteResponse::build(tid, payload.len() as i32);
-            io_thread_data_tx
-                .send(IkcFrame::Message(write_response))
-                .await?;
-
-            sleep(Duration::from_millis(CLEANUP_SLEEP_DURATION)).await;
-
-            pb.inc(1);
-        }
-
-        io_control_command_tx
-            .send(IoControlCommand::Shutdown)
-            .await?;
-        if let Some(response) = io_control_response_rx.recv().await {
-            if response != IoControlResponse::Shutdown {
-                let reason: String =
-                    format!("unexpected control response received during shutdown: {response:?}");
-                error!("run_warm_start_vmm(): {reason}");
-                anyhow::bail!(reason);
-            }
-        } else {
-            let reason: String = "I/O control response channel closed before receiving shutdown \
-                                  acknowledgment"
-                .to_string();
-            error!("run_warm_start_vmm(): {reason}");
-            anyhow::bail!(reason);
-        }
-
-        drop(io_thread_data_tx);
-        drop(io_control_command_tx);
-
-        match user_vm_handle.await {
-            Ok(Ok(exit_status)) => {
-                if exit_status != 0 {
-                    let reason: String =
-                        format!("error running user VM, exit-status={exit_status}");
-                    error!("{reason}");
-                    return Err(anyhow::anyhow!(reason));
-                }
-                debug!("User VM: done running");
-            },
-            Ok(Err(error)) => {
-                error!("error running user VM: {error:?}");
-                return Err(error);
-            },
-            Err(error) => {
-                let reason: String = format!("error joining user VM task: {error:?}");
-                error!("{reason}");
-                return Err(anyhow::anyhow!(reason));
-            },
-        }
-
-        pb.finish();
-        latencies.sort();
-        println!("p50: {} us", latencies[(self.iterations as f32 * 0.5) as usize]);
-        println!("p95: {} us", latencies[(self.iterations as f32 * 0.95) as usize]);
-        println!("p99: {} us", latencies[(self.iterations as f32 * 0.99) as usize]);
-
-        Ok(())
-    }
-
-    #[cfg(feature = "timestamp-messages")]
-    pub async fn run_echo_breakdown(&mut self, linuxd_deployment: &LinuxdDeployment) -> Result<()> {
-        // First start nanvixd and the user VM.
-        let (new_msg_headers, new_msg) = self.prepare_new_message(None, None)?;
-        self.setup(linuxd_deployment);
-        let (user_vm_id, mut gateway_stream) = self
-            .start(new_msg, new_msg_headers, linuxd_deployment)
-            .await?;
-
-        // The labels in this array are also added as comments to the line of code where the
-        // timestamp is added.
-        let steps: Vec<&str> = vec![
-            // In-path
-            "nanvix-bench::write_all()",                    // 0
-            "linuxd::worker_thread::handle_read_request()", // 1
-            "uservm::io_thread::system_vm::read()",         // 2
-            "uservm::memory_thread::data_rx::recv()",       // 3
-            "uservm::lib::vm_input::vmexit()",              // 4
-            "uservm::lib::vm_input::vm_write_bytes()",      // 5
-            // Out-path
-            "uservm::lib::vm_output::send()",                // 6
-            "uservm::io_thread::system_vm::write()",         // 7
-            "linuxd::worker_thread::handle_write_request()", // 8
-            "nanvix-bench::read_exact()",                    // 9
-        ];
-
-        let header_size = 1;
-        let data_size = header_size + profiler::MAX_NUMBER_MESSAGE_TIMESTAMPS * 2;
-
-        // Warmup: send one untimed echo to trigger lazy initialization (worker thread
-        // creation, TCP path warm-up, etc.) so that timed iterations reflect steady-state
-        // latency.
-        {
-            let warmup_data: Vec<u8> = vec![0u8; data_size];
-            let mut warmup_response: Vec<u8> = vec![0u8; data_size];
-            gateway_stream.write_all(&warmup_data).await?;
-            gateway_stream.read_exact(&mut warmup_response).await?;
-            sleep(Duration::from_millis(WARMUP_SLEEP_DURATION)).await;
-        }
-
-        // For each different step we measure, we record the delta for each iteration.
-        let mut latencies: Vec<Vec<u16>> = Vec::with_capacity(steps.len() + 1);
-        for _ in 0..(steps.len() + 1) {
-            latencies.push(vec![0u16; self.iterations]);
-        }
-
-        // Display a progress bar.
-        let pb: ProgressBar = ProgressBar::new(self.iterations.try_into().unwrap());
-        pb.set_style(
-            ProgressStyle::default_bar()
-                .template("{msg} [{bar:40.cyan/blue}] {pos}/{len} ({percent}%)")
-                .expect("error creating progress bar")
-                .progress_chars("#>-"),
-        );
-        pb.set_message("Benchmark progress:");
-
-        for iter in 0..self.iterations {
-            let mut data: Vec<u8> = vec![0u8; data_size];
-            let mut response: Vec<u8> = vec![0u8; data_size];
-
-            // Add initial timestamp
-            // Label: nanvix-bench::write_all()
-            profiler::timestamp_message!(&mut data, 0);
-
-            gateway_stream.write_all(&data).await?;
-            gateway_stream.read_exact(&mut response).await?;
-
-            // Add final timestamp.
-            // Label: nanvix-bench::read_exact()
-            profiler::timestamp_message!(&mut response, 0);
-
-            // Process results.
-            let mut first_timestamp: Option<u16> = None;
-            let mut last_timestamp: Option<u16> = None;
-            let num_stamps: usize = response[0] as usize;
-            if num_stamps != steps.len() {
-                return Err(anyhow::anyhow!(
-                    "not enough timestamps (got={num_stamps}, expected={})",
-                    steps.len()
-                ));
-            }
-            for (step_idx, chunk) in (0..num_stamps).zip(response[header_size..].chunks_exact(2)) {
-                let timestamp: u16 = u16::from_le_bytes([chunk[0], chunk[1]]);
-
-                if first_timestamp.is_none() {
-                    first_timestamp = Some(timestamp);
-                }
-
-                if let Some(last) = last_timestamp {
-                    let delta: u16 = timestamp.wrapping_sub(last);
-                    latencies[step_idx][iter] = delta;
-                }
-
-                last_timestamp = Some(timestamp);
-            }
-
-            if first_timestamp.is_some() && last_timestamp.is_some() {
-                latencies[steps.len()][iter] = last_timestamp.unwrap() - first_timestamp.unwrap()
-            } else {
-                return Err(anyhow::anyhow!("have not collected enough timestamps!"));
-            }
-
-            pb.inc(1);
-        }
-
-        pb.finish();
-
-        // Clean-up.
-        self.kill(user_vm_id).await?;
-        self.cleanup();
-
-        // Print results
-        for step_idx in 0..(steps.len() + 1) {
-            if step_idx < steps.len() {
-                print!("{step_idx:<2} | {:<48}", steps[step_idx]);
-            } else {
-                print!("{step_idx:<2} | {:<48}", "Total");
-            }
-
-            if step_idx == 0 {
-                println!(" | First Step");
-                continue;
-            }
-
-            latencies[step_idx].sort();
-            print!(
-                " | p50: {:5} | p95: {:5} | p99 {:5}",
-                latencies[step_idx][(self.iterations as f32 * 0.5) as usize],
-                latencies[step_idx][(self.iterations as f32 * 0.95) as usize],
-                latencies[step_idx][(self.iterations as f32 * 0.99) as usize],
-            );
-
-            if step_idx < steps.len() && steps[step_idx] == "microvm::mod::vm_input::vmexit()" {
-                println!(" | Time for VM to react to IO being avail.");
-            } else {
-                println!();
-            }
-        }
-
-        Ok(())
-    }
-}
 
 #[tokio::main]
 async fn main() -> Result<()> {
